fix(my-space): #4222 — corrige les retours des audits a11y, sécurité et structure
Trois remédiations, plus l'extraction qu'elles ont rendue nécessaire.

RGAA 6.1 / WCAG 2.4.4 — en variant `closed`, la ligne de l'étape 1 est de
nouveau rendue, si bien que deux boutons œil coexistaient sur la même page avec
le même nom accessible ("Voir le récapitulatif de la déclaration") pour deux
destinations différentes. Les boutons étant en icône seule, ce libellé est leur
seul nom : les lignes de seconde déclaration portent désormais le leur.

Journalisation d'audit — `jointEvaluation.getFile` expose un document
d'entreprise et n'avait jusqu'ici aucun appelant ; la brancher rend le trou de
journalisation réel. Câblée sur les trois points requis, catégorie
`read_sensitive`.

Taille de fichier — `VerticalStepper.tsx` franchissait le seuil de 400 lignes.
Les contenus d'étape partent dans `StepContent.tsx`, les primitives de ligne
dans `StepRows.tsx` ; `JointEvaluationFileInfo` reste ré-exporté pour ne pas
toucher ses trois consommateurs. Déplacement pur, aucun changement de
comportement.
